backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0
